requirements.txt
bidict==0.23.1
blinker==1.7.0
Brotli==1.1.0
cachelib==0.9.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Compress==1.14
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.5.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==8.3.0
ordered-set==4.1.0
packaging==23.2
Pygments==2.17.2
PyMySQL==1.1.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
rich==13.7.0
simple-websocket==1.0.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
zipp==3.17.0
